Reorder selected tested modules and functions
