+++ tests/helpers/degradation_results.py
@@ -1,5 +1,3 @@
-import numpy as np
-
 import perun.check.integral_comparison as int_cmp
 import perun.check.linear_regression as lreg
 import perun.check.local_statistics as loc_stat
@@ -13,7 +11,7 @@ _PREG_EXPECTED_RESULTS = [
     [
         # CONSTANT
         [
-            # ERROR 
+            # ERROR
             {'result': pc.Degradation, 'type': 'constant', 'rate': 999},
             # IMPROVEMENT
             {'result': pc.Optimization, 'type': 'constant', 'rate': -91},
