+++ tests/test_regression_analysis.py
@@ -12,37 +12,9 @@ import pytest
 import perun.utils.exceptions as exceptions
 from perun.postprocess.regression_analysis.run import postprocess
 import tests.helpers.utils as test_utils
+from tests.helpers.utils import compare_results, generate_models_by_uid
 
 __author__ = 'Jiri Pavela'
-
-
-def compare_results(expected, actual, eps=0.0001):
-    """Compare two float values with eps tolerance.
-
-    Arguments:
-        expected(float): the expected result value
-        actual(float): the actual result value
-        eps(float): the tolerance value
-    Returns:
-        None
-    """
-    assert abs(abs(expected) - abs(actual)) < eps
-
-
-def generate_models_by_uid(profile, value, uid_sequence, key='model'):
-    """Provides computed models results for each uid in the specified uid sequence.
-
-    Arguments:
-        profile(Profile): the whole profile with 'models' results
-        value(str): the specification of value of given key for matching models
-        uid_sequence(list of str): list of uid values to search for
-        key(str): the key for matching models
-    Returns:
-        generator: stream of lists with models dictionaries according to uid sequence
-    """
-    models = profile['profile']['models']
-    for uid in uid_sequence:
-        yield [m for m in models if m['uid'] == uid and m[key] == value]
 
 
 def test_incorrect_calls(postprocess_profiles):
